CC/chatbot-api/requirements.txt
Flask==3.0.0
pandas==2.0.3
numpy==1.26.0
gunicorn==21.2.0
google-cloud-storage==2.14.0

